Refactor Marketplace service to add permissions
