+++ src/main/java/org/craftercms/studio/impl/v2/service/marketplace/MarketplaceServiceImpl.java
@@ -17,17 +17,12 @@
 
 package org.craftercms.studio.impl.v2.service.marketplace;
 
-import java.io.IOException;
 import java.util.Map;
 
-import org.apache.commons.lang3.StringUtils;
-import org.craftercms.commons.monitoring.VersionInfo;
-import org.craftercms.commons.plugin.model.Plugin;
-import org.craftercms.commons.plugin.model.PluginDescriptor;
-import org.craftercms.commons.plugin.model.Version;
-import org.craftercms.commons.rest.RestTemplate;
-import org.craftercms.studio.api.v1.constant.StudioConstants;
-import org.craftercms.studio.api.v1.exception.BlueprintNotFoundException;
+import org.craftercms.commons.security.permissions.DefaultPermission;
+import org.craftercms.commons.security.permissions.annotations.HasPermission;
+import org.craftercms.commons.validation.annotations.param.ValidateParams;
+import org.craftercms.commons.validation.annotations.param.ValidateStringParam;
 import org.craftercms.studio.api.v1.exception.ServiceLayerException;
 import org.craftercms.studio.api.v1.exception.repository.InvalidRemoteRepositoryCredentialsException;
 import org.craftercms.studio.api.v1.exception.repository.InvalidRemoteRepositoryException;
@@ -36,30 +31,10 @@ import org.craftercms.studio.api.v1.exception.repository.RemoteRepositoryNotBare
 import org.craftercms.studio.api.v1.exception.repository.RemoteRepositoryNotFoundException;
 import org.craftercms.studio.api.v1.log.Logger;
 import org.craftercms.studio.api.v1.log.LoggerFactory;
-import org.craftercms.studio.api.v1.service.site.SiteService;
-import org.craftercms.studio.api.v2.dal.RemoteRepository;
 import org.craftercms.studio.api.v2.exception.marketplace.MarketplaceException;
-import org.craftercms.studio.api.v2.exception.marketplace.MarketplaceUnreachableException;
-import org.craftercms.studio.api.v2.exception.marketplace.MarketplaceNotInitializedException;
 import org.craftercms.studio.api.v2.service.marketplace.MarketplaceService;
-import org.craftercms.studio.api.v2.service.marketplace.Constants;
-import org.craftercms.studio.api.v2.service.marketplace.Paths;
-import org.craftercms.studio.api.v2.service.site.internal.SitesServiceInternal;
-import org.craftercms.studio.api.v2.service.system.InstanceService;
-import org.craftercms.studio.api.v2.utils.StudioConfiguration;
+import org.craftercms.studio.api.v2.service.marketplace.internal.MarketplaceServiceInternal;
 import org.craftercms.studio.model.rest.marketplace.CreateSiteRequest;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.client.ResourceAccessException;
-import org.springframework.web.util.UriComponentsBuilder;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import static org.craftercms.studio.api.v2.service.marketplace.Constants.PLUGIN_REF;
-import static org.craftercms.studio.api.v2.service.marketplace.Constants.PLUGIN_URL;
 
 /**
  * Default implementation of {@link MarketplaceService} that proxies all request to the configured Marketplace
@@ -67,166 +42,33 @@ import static org.craftercms.studio.api.v2.service.marketplace.Constants.PLUGIN_
  * @author joseross
  * @since 3.1.2
  */
-public class MarketplaceServiceImpl implements MarketplaceService, InitializingBean {
+public class MarketplaceServiceImpl implements MarketplaceService {
 
     private static final Logger logger = LoggerFactory.getLogger(MarketplaceServiceImpl.class);
 
-    public MarketplaceServiceImpl(final InstanceService instanceService, final SiteService siteService,
-                                  final SitesServiceInternal sitesServiceInternal,
-                                  final StudioConfiguration studioConfiguration) {
-        this.instanceService = instanceService;
-        this.siteService = siteService;
-        this.sitesServiceInternal = sitesServiceInternal;
-        this.studioConfiguration = studioConfiguration;
+    public MarketplaceServiceImpl(final MarketplaceServiceInternal marketplaceServiceInternal) {
+        this.marketplaceServiceInternal = marketplaceServiceInternal;
     }
 
-    protected InstanceService instanceService;
+    protected MarketplaceServiceInternal marketplaceServiceInternal;
 
-    protected SiteService siteService;
-
-    protected SitesServiceInternal sitesServiceInternal;
-
-    protected StudioConfiguration studioConfiguration;
-
-    protected RestTemplate restTemplate = new RestTemplate();
-
-    protected ObjectMapper mapper = new ObjectMapper();
-
-    /**
-     * The custom HTTP headers to sent with all requests
-     */
-    protected HttpHeaders httpHeaders;
-
-    /**
-     * The current Crafter CMS version, sent with all requests
-     */
-    protected String version;
-
-    /**
-     * The current Crafter CMS edition, sent with all requests
-     */
-    protected String edition;
-
-    /**
-     * The Marketplace URL to use
-     */
-    protected String url;
-
-    public void setUrl(final String url) {
-        this.url = url;
-    }
 
     @Override
-    public void afterPropertiesSet() throws IOException {
-        VersionInfo versionInfo = VersionInfo.getVersion(MarketplaceServiceImpl.class);
-        if (versionInfo == null) {
-            logger.warn("Marketplace service could not be initialized");
-            return;
-        }
-        String versionStr = versionInfo.getPackageVersion();
-
-        // init version
-        version = Version.getVersion(versionStr);
-        edition = Version.getEdition(versionStr);
-
-        // init headers
-        httpHeaders = new HttpHeaders();
-        httpHeaders.set(HEADER_STUDIO_ID, instanceService.getInstanceId());
-
-        httpHeaders.set(HEADER_STUDIO_BUILD, versionInfo.getPackageBuild());
-        httpHeaders.set(HEADER_STUDIO_VERSION, versionInfo.getPackageVersion());
-        httpHeaders.set(HEADER_JAVA_VERSION, versionStr);
-
-        httpHeaders.set(HEADER_OS_NAME, versionInfo.getOsName());
-        httpHeaders.set(HEADER_OS_VERSION, versionInfo.getOsVersion());
-        httpHeaders.set(HEADER_OS_ARCH, versionInfo.getOsArch());
-    }
-
-    @Override
-    public Map<String, Object> searchPlugins(final String type, final String keywords, final long offset,
-                                             final long limit)
+    @ValidateParams
+    @HasPermission(type = DefaultPermission.class, action = "create-site")
+    public Map<String, Object> searchPlugins(@ValidateStringParam(name = "type") String type,
+                                             @ValidateStringParam(name = "keywords") String keywords,
+                                             long offset, long limit)
         throws MarketplaceException {
-
-        validate();
-        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(url)
-            .path(Paths.PLUGIN_SEARCH)
-            .queryParam(Constants.PARAM_VERSION, version)
-            .queryParam(Constants.PARAM_EDITION, edition)
-            .queryParam(Constants.PARAM_OFFSET, offset)
-            .queryParam(Constants.PARAM_LIMIT, limit);
-
-        if (StringUtils.isNotEmpty(type)) {
-            builder.queryParam(Constants.PARAM_TYPE, type);
-        }
-
-        if (StringUtils.isNotEmpty(keywords)) {
-            builder.queryParam(Constants.PARAM_KEYWORDS, keywords);
-        }
-
-        HttpEntity<Void> request = new HttpEntity<>(null, httpHeaders);
-
-        try {
-            ResponseEntity<Map<String, Object>> response =
-                restTemplate.exchange(builder.build().toString(), HttpMethod.GET, request,
-                    new ParameterizedTypeReference<Map<String, Object>>() {});
-            return response.getBody();
-        } catch (ResourceAccessException e) {
-            throw new MarketplaceUnreachableException(url, e);
-        }
-    }
-
-    protected void validate() throws MarketplaceException {
-        if (StringUtils.isEmpty(version)) {
-            throw new MarketplaceNotInitializedException();
-        }
-    }
-
-    public Map<String, Object> getDescriptor(String id, Version version) throws MarketplaceException,
-        BlueprintNotFoundException {
-
-        validate();
-        UriComponentsBuilder builder = UriComponentsBuilder.fromHttpUrl(url)
-            .path(Paths.GET_PLUGIN)
-            .pathSegment(id, String.format("%s.%s.%s", version.getMajor(), version.getMinor(), version.getPatch()));
-
-        HttpEntity<Void> request = new HttpEntity<>(null, httpHeaders);
-
-        try {
-            ResponseEntity<Map<String, Object>> response =
-                restTemplate.exchange(builder.build().toString(), HttpMethod.GET, request,
-                    new ParameterizedTypeReference<Map<String, Object>>() {});
-            return response.getBody();
-        } catch (ResourceAccessException e) {
-            throw new MarketplaceUnreachableException(url, e);
-        } catch (IllegalArgumentException e) {
-            throw new BlueprintNotFoundException(String.format("Blueprint not found in the Marketplace: %s %s.%s.%s",
-             id, version.getMajor(), version.getMinor(), version.getPatch()));
-        }
+        return marketplaceServiceInternal.searchPlugins(type, keywords, offset, limit);
     }
 
     @Override
+    @HasPermission(type = DefaultPermission.class, action = "create-site")
     public void createSite(CreateSiteRequest request) throws RemoteRepositoryNotFoundException,
         InvalidRemoteRepositoryException, RemoteRepositoryNotBareException, InvalidRemoteUrlException,
         ServiceLayerException, InvalidRemoteRepositoryCredentialsException {
-
-        if (StringUtils.isEmpty(request.getSandboxBranch())) {
-            request.setSandboxBranch(studioConfiguration.getProperty(StudioConfiguration.REPO_SANDBOX_BRANCH));
-        }
-
-        if (StringUtils.isEmpty(request.getRemoteName())) {
-            request.setRemoteName(studioConfiguration.getProperty(StudioConfiguration.REPO_DEFAULT_REMOTE_NAME));
-        }
-
-        Map<String, Object> result = getDescriptor(request.getBlueprintId(), request.getBlueprintVersion());
-        Plugin plugin = mapper.convertValue(result, Plugin.class);
-
-        sitesServiceInternal.validateBlueprintParameters(PluginDescriptor.of(plugin), request.getParameters());
-
-        siteService.createSiteWithRemoteOption(request.getSiteId(), request.getSandboxBranch(),
-            request.getDescription(), request.getBlueprintId(), request.getRemoteName(),
-            result.get(PLUGIN_URL).toString(), result.get(PLUGIN_REF).toString(), false,
-            RemoteRepository.AuthenticationType.NONE, null, null, null,
-            null, StudioConstants.REMOTE_REPOSITORY_CREATE_OPTION_CLONE, request.getParameters(),
-            true);
+        marketplaceServiceInternal.createSite(request);
     }
+
 }
